Add toolbar modes: simple and advanced
In the simple mode, which is displayed by default, there are only
a few buttons, along with the modeswitch button and preview button.

The advanced or full mode contains all the toolbar buttons.

On clicking the modeswitch button, the toolbar is expanded to show
all the buttons, and the modeswitch button becomes a 'close' button.
If this is clicked, the toolbar reverts to the simple view.