Add Live Preview
Added a right sidebar that continually updates according to the
content in the editor. It displays a small preview image, which when
clicked will open the full-size preview.

Removed the Preview button from the toolbar as it is now redundant.

Added a 'Post to Twitter' button on top of the right sidebar. For
simple posts, this should allow the user to finish everything on
a single page, with just one click.